requirements.txt
absl-py==0.11.0
alembic==1.4.3
anndata==0.7.8
appnope==0.1.2
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
banal==1.0.1
bayesian-optimization==1.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
clang==5.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.6
dataset==1.4.4
debugpy==1.5.0
decorator==4.4.2
Deprecated==1.2.13
entrypoints==0.3
fasteners==0.16.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.13.0
glfw==2.1.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.39.0
gym==0.19.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-resources==5.4.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==0.17.0
jsonschema==4.4.0
jupyter-client==7.0.6
jupyter-core==4.8.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.38.0
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
msgpack==1.0.3
mujoco-py==2.0.2.8
natsort==8.1.0
nest-asyncio==1.5.1
networkx==2.5
numba==0.55.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.4
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip-chill==1.0.1
prompt-toolkit==3.0.20
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pynndescent==0.5.6
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.4.1
pyzmq==22.3.0
ray==1.9.2
redis==4.1.2
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.45
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
sinfo==0.3.4
six==1.15.0
sklearn==0.0
snakeviz==2.1.1
SQLAlchemy==1.3.20
statsmodels==0.13.2
stdlib-list==0.8.0
tables==3.7.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.9.0
torch-tb-profiler==0.3.1
torchvision==0.10.0
tornado==6.1
tqdm==4.55.1
traitlets==5.1.0
transformers==4.1.1
typing-extensions==3.7.4.3
umap-learn==0.5.2
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
xlrd==1.2.0
zipp==3.7.0
